+++ app/src/main/java/org/fossasia/openevent/api/processor/EventListResponseProcessor.java
@@ -2,10 +2,14 @@ package org.fossasia.openevent.api.processor;
 
 import android.util.Log;
 
+import com.squareup.otto.Bus;
+
+import org.fossasia.openevent.OpenEventApp;
 import org.fossasia.openevent.api.protocol.EventResponseList;
 import org.fossasia.openevent.data.Event;
 import org.fossasia.openevent.dbutils.DbContract;
 import org.fossasia.openevent.dbutils.DbSingleton;
+import org.fossasia.openevent.events.FailedDownload;
 
 import java.util.ArrayList;
 
@@ -38,6 +42,7 @@ public class EventListResponseProcessor implements Callback<EventResponseList> {
 
     @Override
     public void failure(RetrofitError error) {
-
+        Bus bus = OpenEventApp.getEventBus();
+        bus.post(new FailedDownload());
     }
 }
