+++ app/src/main/java/org/fossasia/openevent/api/processor/MicrolocationListResponseProcessor.java
@@ -2,10 +2,14 @@ package org.fossasia.openevent.api.processor;
 
 import android.util.Log;
 
+import com.squareup.otto.Bus;
+
+import org.fossasia.openevent.OpenEventApp;
 import org.fossasia.openevent.api.protocol.MicrolocationResponseList;
 import org.fossasia.openevent.data.Microlocation;
 import org.fossasia.openevent.dbutils.DbContract;
 import org.fossasia.openevent.dbutils.DbSingleton;
+import org.fossasia.openevent.events.FailedDownload;
 
 import java.util.ArrayList;
 
@@ -37,6 +41,7 @@ public class MicrolocationListResponseProcessor implements Callback<Microlocatio
 
     @Override
     public void failure(RetrofitError error) {
-
+        Bus bus = OpenEventApp.getEventBus();
+        bus.post(new FailedDownload());
     }
 }
