+++ app/src/main/java/org/fossasia/openevent/dbutils/DbSingleton.java
@@ -90,6 +90,10 @@ public class DbSingleton {
         return databaseOperations.getVersionIds(mDb);
     }
 
+    public void addBookmarks(int id) {
+        databaseOperations.addToDb(id);
+    }
+
     public Speaker getSpeakerById(int id) {
         getReadOnlyDatabase();
         return databaseOperations.getSpeakerById(id, mDb);
@@ -116,7 +120,7 @@ public class DbSingleton {
     }
 
     public ArrayList<Speaker> getSpeakersbySessionName(String sessionName) throws ParseException {
-        return databaseOperations.getSpeakersbySessionname(sessionName, mDb);
+        return databaseOperations.getSpeakersbySessionName(sessionName, mDb);
     }
