+++ app/src/main/java/org/fossasia/openevent/events/EventDownloadEvent.java
@@ -1,0 +1,7 @@
+package org.fossasia.openevent.events;
+
+/**
+ * Created by MananWason on 8/5/2015.
+ */
+public class EventDownloadEvent {
+}
