+++ app/src/main/java/org/fossasia/openevent/events/SpeakerDownloadEvent.java
@@ -1,0 +1,7 @@
+package org.fossasia.openevent.events;
+
+/**
+ * Created by MananWason on 8/4/2015.
+ */
+public class SpeakerDownloadEvent {
+}
