+++ app/src/main/res/layout/activity_main.xml
@@ -2,9 +2,9 @@
 <FrameLayout
     android:id="@+id/layout_main"
     xmlns:android="http://schemas.android.com/apk/res/android"
-             xmlns:app="http://schemas.android.com/apk/res-auto"
-             android:layout_width="match_parent"
-             android:layout_height="match_parent">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -27,6 +27,7 @@
             android:layout_marginTop="-7dp"
             android:max="100"
             android:visibility="gone"/>
+
         <android.support.v4.widget.DrawerLayout
             android:id="@+id/drawer"
             android:layout_width="match_parent"
